api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.5
elasticsearch==8.14.0
python-dateutil==2.9.0.post0
pydantic==2.8.2
